feat: add a DataTrailEmbedded to @grafana/runtime
