+++ public/app/features/trails/Integrations/DataTrailEmbedded.tsx
@@ -1,16 +1,12 @@
 import React from 'react';
 
-import { AdHocVariableFilter } from '@grafana/data';
-import { SceneComponentProps, SceneObjectBase, SceneObjectState, SceneTimeRangeLike } from '@grafana/scenes';
+import { DataTrailEmbeddedState as DataTrailEmbeddedStateRuntime } from '@grafana/runtime';
+import { SceneComponentProps, SceneObjectBase } from '@grafana/scenes';
 
 import { DataTrail } from '../DataTrail';
 
-export interface DataTrailEmbeddedState extends SceneObjectState {
-  timeRange: SceneTimeRangeLike;
-  metric?: string;
-  filters?: AdHocVariableFilter[];
-  dataSourceUid?: string;
-}
+export type DataTrailEmbeddedState = DataTrailEmbeddedStateRuntime;
+
 export class DataTrailEmbedded extends SceneObjectBase<DataTrailEmbeddedState> {
   static Component = DataTrailEmbeddedRenderer;
